+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## 3.3.0-TERUMO - 2024-06-12
 ### Changed
 - YLP-3075 As a user, I want to be able to choose Japan or UK country
+- YLP-3087 As a user, I want to see the medisafe pump event in yourloops
 
 ### Engineering use
 - YLP-2948 Updated failing tests because of the time change